# remove simd, improves perf on arm64 cpu by x2