--- a/src/functional.rs
+++ b/src/functional.rs
@@ -215,10 +215,9 @@ pub fn matmul_q8(xout: &mut [f32], x: &MutableQuantizedTensor, w: &QuantizedTens
 
 pub fn matmul_q4(xout: &mut [f32], x: &MutableQuantizedTensor, w: &QuantizedTensor, n: usize, o: usize, gs: usize) {
     let group_size = gs / 2;
-    let n_simd = group_size / 8;
 
-    let mask_a = i32x8::new([0x0F; 8]);
-    let mask_b = i32x8::new([0xF0; 8]);
+    let mask_a: i16 = 0x0F;
+    let mask_b: i16 = 0xF0;
     
     xout.par_chunks_exact_mut(o).enumerate().for_each(|(j, elem)| {
         let xi = j*n;
@@ -227,23 +226,22 @@ pub fn matmul_q4(xout: &mut [f32], x: &MutableQuantizedTensor, w: &QuantizedTens
             let ni: usize = i * n / 2;
 
             *xout_elem = (0..=(n/2 - group_size)).step_by(group_size).map(|j| {
-                let mut ival = i32x8::ZERO;
+                let mut ival = 0;
 
-                for k in 0..n_simd {
-                    let x_vec = i32x8::from(&x.q[xi+j+k*8..xi+j+k*8+8]);
-                    let w_vec = i32x8::from(&w.q[ni+j+k*8..ni+j+k*8+8]);
+                let x_vec = &x.q[xi+j..xi+j+group_size];
+                let w_vec = &w.q[ni+j..ni+j+group_size];
 
-                    let x_a = (x_vec & mask_a) - 8;
-                    let w_a = (w_vec & mask_a) - 8;
-                    
-                    let x_b = (mask_a & ((x_vec & mask_b) >> 4)) - 8;
-                    let w_b = (mask_a & ((w_vec & mask_b) >> 4)) - 8;
+                for l in 0..group_size {
+                    let x_a = (x_vec[l] as i16 & mask_a) - 8;
+                    let w_a = (w_vec[l] as i16 & mask_a) - 8;
 
-                    ival += x_a * w_a;
-                    ival += x_b * w_b;
+                    let x_b = ((x_vec[l] as i16 & mask_b) >> 4) - 8;
+                    let w_b = ((w_vec[l] as i16 & mask_b) >> 4) - 8;
+
+                    ival += x_a * w_a + x_b * w_b;
                 }
 
-                (ival.reduce_add() as f32) * w.s[(ni + j) / group_size] * x.s[(xi + j) / group_size] 
+                (ival as f32) * w.s[(ni + j) / group_size] * x.s[(xi + j) / group_size] 
             }).sum();
         });
     });