On Arm64 cpu, removing the SIMD in matmul allows the compiler to optimise it for arm64. The performance in my experiments goes from around ~2 t/s to ~4.5 t/s.

 I'm not sure if this has any performance implications for running on x86, as I don't have one to test on. If it does make it slower, than I'm happy to put this version behind a target flag for arm64 e.g. https://docs.rs/core_arch/latest/core_arch/#static-cpu-feature-detection

Before
```
❯ cargo build --release --bin chat && ./target/release/chat --model model/llama3.2-3b-it-q40.lmrs --tokenizer model/tokenizer.bin --show-metrics --temperature 0
   Compiling lmrs v0.1.0 (../lm.rs)
    Finished `release` profile [optimized] target(s) in 22.27s

    L      M     M  RRRR    ssss
    L      MM   MM  R   R  s
    L      M M M M  RRRR    sss
    L      M  M  M  R  R       s
    LLLL   M     M  R   R  sssss
    
LMRS version: 4
Model type: LLAMA

Using Q4_0 quantization.
Loading weights...
Done.

You: hello there
Assistant:
Hello! It's nice to meet you. Is there something I can help you with, or would you like to chat for a bit?
Speed: 1.98 tok/s
You: 
```

After
```
❯ cargo build --release --bin chat && ./target/release/chat --model model/llama3.2-3b-it-q40.lmrs --tokenizer model/tokenizer.bin --show-metrics --temperature 0
   Compiling lmrs v0.1.0 (../lm.rs)
    Finished `release` profile [optimized] target(s) in 21.72s

    L      M     M  RRRR    ssss
    L      MM   MM  R   R  s
    L      M M M M  RRRR    sss
    L      M  M  M  R  R       s
    LLLL   M     M  R   R  sssss
    
LMRS version: 4
Model type: LLAMA

Using Q4_0 quantization.
Loading weights...
Done.

You: hello there
Assistant:
Hello! It's nice to meet you. Is there something I can help you with, or would you like to chat for a bit?
Speed: 4.67 tok/s
You:
```