@samuel-vitorino 2024-10-16
Good morning. Thank you for the feedback. Does this happen on the 8-bit quantization models too? On x86, SIMD helps improve performance significantly on Q8. Anyways, Q4 is still being improved and I will test today on my x86 machine if I get the same results.