@smpurkis 2024-10-16
Good question on Q8, I'll see if it makes a similar difference 👍