@smpurkis 2024-10-16
On Q8 it gives a smaller improvement. 2.90 t/s -> 3.70 t/s.