@samuel-vitorino 2024-10-16
Without having other people testing on arm64 systems it's hard to say if it's justified adding a target flag for the Q4 and Q8 matmuls. The reason I used the wide crate was that it falls back to normal instructions when SIMD isn't available, so I'm not sure what's causing the slowdown.