@samuel-vitorino 2024-10-16
Just to be sure, are you using the ``RUSTFLAGS="-C target-cpu=native"`` flag?