@smpurkis 2024-10-16
I wasn't, however rerunning all of the above on `main` and on my branch doesn't change the tokens per second